requirements.txt
streamlit
langchain
langchain-community
langchain-core
faiss-cpu
langchain-google-genai
sentence-transformers
pypdf
beautifulsoup4
selenium
python-dotenv
langchain-groq
langchain-ollama
langchain-huggingface
watchdog
langchain-text-splitters
unstructured
lxml
